requirements.txt
fastapi
numpy
uvicorn
pydantic
pymongo
bcrypt
pyjwt
torch
torchvision
tensorflow